# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TernakLomba API Documentation",
        "description": "The **TernakLomba API** is a robust and scalable solution for managing competitive events, participants, scores, and results. \n *   It is designed to support a wide range of competitions, from small local events to large-scale tournaments. \n *  Key features include:\n *  - **Competition Management**: Create, update, and manage competitions with details such as name, description, start/end dates, and status.\n *  - **Participant Management**: Register and manage participants, including their details and competition assignments.\n *  - **Score Management**: Submit and manage scores for participants, with support for multiple judges and scoring criteria.\n *  - **Results and Leaderboards**: Automatically calculate and display competition results, including rankings and leaderboards.\n *  - **Judging Criteria**: Define custom judging criteria for competitions to ensure fair and transparent evaluations.\n *  - **Notifications**: Send real-time notifications to participants and judges about competition updates, scores, and results.\n *  This API is ideal for event organizers, educational institutions, and organizations hosting competitive events.",
        "contact": {
            "name": "the developer",
            "email": "support@ternaklomba.com"
        },
        "license": {
            "name": "Apache 2.0",
...
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/lombaHadiah": {
            "get": {
                "tags": [
                    "Lomba Hadiah"
                ],
                "summary": "Get All Lomba Hadiah",
                "description": "Mengambil Semua data Lomba Hadiah",
                "operationId": "lombaHadiah-all",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengambil data lomba hadiah",
                                    "data": [
                                        {
                                            "lomba_id": 1,
                                            "typeHadiah_id": 1,
                                            "quantity": 1
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lomba Hadiah"
                ],
                "summary": "Create Lomba Hadiah",
                "description": "Membuat Lomba Hadiah",
                "operationId": "lombaHadiah-store",
                "parameters": [
                    {
                        "name": "lomba_id",
                        "in": "query",
                        "description": "Lomba ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "typeHadiah_id",
                        "in": "query",
                        "description": "Type Hadiah ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Jumlah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil menambahkan data lomba hadiah",
                                    "data": {
                                        "lomba_id": 1,
                                        "typeHadiah_id": 1,
                                        "quantity": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lombaHadiah/get/{id_hadiah}": {
            "get": {
                "tags": [
                    "Lomba Hadiah"
                ],
                "summary": "Get One Lomba Hadiah",
                "description": "Mengambil data Lomba Hadiah",
                "operationId": "lombaHadiah-get",
                "parameters": [
                    {
                        "name": "id_hadiah",
                        "in": "path",
                        "description": "Lomba Hadiah ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengambil data lomba hadiah",
                                    "data": {
                                        "lomba_id": 1,
                                        "typeHadiah_id": 1,
                                        "quantity": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lombaHadiah/{id_hadiah}": {
            "put": {
                "tags": [
                    "Lomba Hadiah"
                ],
                "summary": "Update Lomba Hadiah",
                "description": "Mengubah data Lomba Hadiah",
                "operationId": "lombaHadiah-update",
                "parameters": [
                    {
                        "name": "id_hadiah",
                        "in": "path",
                        "description": "Lomba Hadiah ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "typeHadiah_id",
                        "in": "query",
                        "description": "Type Hadiah ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": 1
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Jumlah",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengubah data lomba hadiah",
                                    "data": {
                                        "lomba_id": 1,
                                        "typeHadiah_id": 1,
                                        "quantity": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lomba Hadiah"
                ],
                "summary": "Delete Lomba Hadiah",
                "description": "Menghapus data Lomba Hadiah",
                "operationId": "lombaHadiah-delete",
                "parameters": [
                    {
                        "name": "id_hadiah",
                        "in": "path",
                        "description": "Lomba Hadiah ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil menghapus data lomba hadiah"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}